pyproject.toml
[tool.poetry]
name = "langchain-sandbox"
version = "0.1.0"
description = ""
authors = ["daichi_yamaoka <tnychor06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.11.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.5"
langchain-openai = "^0.0.5"
numpy = "^1.26.3"
chainlit = "^1.0.200"
spacy = "^3.7.2"
pymupdf = "^1.23.21"
chromadb = "^0.4.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
